+++ pkg/oauth/config.go
@@ -31,8 +31,9 @@ type StackConfig struct {
 
 // FrontendConfig is the configuration for the OAuth frontend.
 type FrontendConfig struct {
-	Language    string `json:"language" name:"-"`
-	StackConfig `json:"stack_config" name:",squash"`
+	Language               string `json:"language" name:"-"`
+	StackConfig            `json:"stack_config" name:",squash"`
+	EnableUserRegistration bool `json:"enable_user_registration" name:"-"`
 }
 
 // Config is the configuration for the OAuth server.