+++ pkg/webui/lib/selectors/env.js
@@ -52,4 +52,6 @@ export const selectSupportLinkConfig = () => selectApplicationConfig().support_l
 
 export const selectDocumentationUrlConfig = () => selectApplicationConfig().documentation_base_url
 
+export const selectEnableUserRegistration = () => selectApplicationConfig().enable_user_registration
+
 export const selectPageData = () => configSelector().PAGE_DATA